+++ client.go
@@ -17,7 +17,7 @@ var pusherPathRegex = regexp.MustCompile("^/apps/([0-9]+)$")
 var maxTriggerableChannels = 100
 
 const (
-	libraryVersion = "5.0.0"
+	libraryVersion = "5.1.0"
 	libraryName    = "pusher-http-go"
 )
 
@@ -142,7 +142,7 @@ be marshallable into JSON.
 
 */
 func (c *Client) Trigger(channel string, eventName string, data interface{}) error {
-	_, err := c.trigger([]string{channel}, eventName, data, TriggerParams{})
+	_, err := c.validateChannelsAndTrigger([]string{channel}, eventName, data, TriggerParams{})
 	return err
 }
 
@@ -195,7 +195,7 @@ func (c *Client) TriggerWithParams(
 	data interface{},
 	params TriggerParams,
 ) (*TriggerChannelsList, error) {
-	return c.trigger([]string{channel}, eventName, data, params)
+	return c.validateChannelsAndTrigger([]string{channel}, eventName, data, params)
 }
 
 /*
@@ -204,7 +204,7 @@ TriggerMulti is the same as `client.Trigger`, except one passes in a slice of
 	client.TriggerMulti([]string{"a_channel", "another_channel"}, "event", data)
 */
 func (c *Client) TriggerMulti(channels []string, eventName string, data interface{}) error {
-	_, err := c.trigger(channels, eventName, data, TriggerParams{})
+	_, err := c.validateChannelsAndTrigger(channels, eventName, data, TriggerParams{})
 	return err
 }
 
@@ -219,7 +219,7 @@ func (c *Client) TriggerMultiWithParams(
 	data interface{},
 	params TriggerParams,
 ) (*TriggerChannelsList, error) {
-	return c.trigger(channels, eventName, data, params)
+	return c.validateChannelsAndTrigger(channels, eventName, data, params)
 }
 
 /*
@@ -232,7 +232,7 @@ Deprecated: use TriggerWithParams instead.
 */
 func (c *Client) TriggerExclusive(channel string, eventName string, data interface{}, socketID string) error {
 	params := TriggerParams{SocketID: &socketID}
-	_, err := c.trigger([]string{channel}, eventName, data, params)
+	_, err := c.validateChannelsAndTrigger([]string{channel}, eventName, data, params)
 	return err
 }
 
@@ -246,18 +246,37 @@ Deprecated: use TriggerMultiWithParams instead.
 */
 func (c *Client) TriggerMultiExclusive(channels []string, eventName string, data interface{}, socketID string) error {
 	params := TriggerParams{SocketID: &socketID}
-	_, err := c.trigger(channels, eventName, data, params)
+	_, err := c.validateChannelsAndTrigger(channels, eventName, data, params)
 	return err
 }
 
-func (c *Client) trigger(channels []string, eventName string, data interface{}, params TriggerParams) (*TriggerChannelsList, error) {
+/*
+SendToUser triggers an event to a specific user.
+Pass in the user id, the event's name, and a data payload. The data payload must
+be marshallable into JSON.
+
+	data := map[string]string{"hello": "world"}
+	client.SendToUser("user123", "say_hello", data)
+*/
+func (c *Client) SendToUser(userId string, eventName string, data interface{}) error {
+	if !validUserId(userId) {
+		return fmt.Errorf("User id '%s' is invalid", userId)
+	}
+	_, err := c.trigger([]string{"#server-to-user-" + userId}, eventName, data, TriggerParams{})
+	return err
+}
+
+func (c *Client) validateChannelsAndTrigger(channels []string, eventName string, data interface{}, params TriggerParams) (*TriggerChannelsList, error) {
 	if len(channels) > maxTriggerableChannels {
 		return nil, fmt.Errorf("You cannot trigger on more than %d channels at once", maxTriggerableChannels)
 	}
 	if !channelsAreValid(channels) {
 		return nil, errors.New("At least one of your channels' names are invalid")
 	}
+	return c.trigger(channels, eventName, data, params)
+}
 
+func (c *Client) trigger(channels []string, eventName string, data interface{}, params TriggerParams) (*TriggerChannelsList, error) {
 	hasEncryptedChannel := false
 	for _, channel := range channels {
 		if isEncryptedChannel(channel) {
@@ -272,6 +291,7 @@ func (c *Client) trigger(channels []string, eventName string, data interface{}, 
 	if hasEncryptedChannel && keyErr != nil {
 		return nil, keyErr
 	}
+
 	if err := validateSocketID(params.SocketID); err != nil {
 		return nil, err
 	}
@@ -462,13 +482,71 @@ func (c *Client) GetChannelUsers(name string) (*Users, error) {
 }
 
 /*
-AuthenticatePrivateChannel allows you to authenticate a users subscription to a
-private channel. It returns authentication signature to send back to the client
-and authorize them.
+AuthenticateUser allows you to authenticate a user s subscription to a
+private channel. It returns an authentication signature to send back to the client
+and authenticate them. In order to identify a user, this method acceps a map containing
+arbitrary user data. It must contain at least and id field with the user's id as a string.
 
 For more information see our docs: http://pusher.com/docs/authenticating_users.
 
-This is an example of authenticating a private-channel, using the built-in
+This is an example of authenticating a user, using the built-in
+Golang HTTP library to start a server.
+
+In order to authenticate a client, one must read the response into type `[]byte`
+and pass it in. This will return a signature in the form of a `[]byte` for you
+to send back to the client.
+
+	func pusherUserAuth(res http.ResponseWriter, req *http.Request) {
+
+		params, _ := ioutil.ReadAll(req.Body)
+		response, err := client.AuthenticateUser(params)
+		if err != nil {
+			panic(err)
+		}
+
+		fmt.Fprintf(res, string(response))
+	}
+
+	func main() {
+		http.HandleFunc("/pusher/user-auth", pusherUserAuth)
+		http.ListenAndServe(":5000", nil)
+	}
+*/
+func (c *Client) AuthenticateUser(params []byte, userData map[string]interface{}) (response []byte, err error) {
+	socketID, err := parseUserAuthenticationRequestParams(params)
+	if err != nil {
+		return
+	}
+
+	if err = validateSocketID(&socketID); err != nil {
+		return
+	}
+
+	if err = validateUserData(userData); err != nil {
+		return
+	}
+
+	var jsonUserData string
+	if jsonUserData, err = jsonMarshalToString(userData); err != nil {
+		return
+	}
+	stringToSign := strings.Join([]string{socketID, "user", jsonUserData}, "::")
+
+	_response := createAuthMap(c.Key, c.Secret, stringToSign, "")
+	_response["user_data"] = jsonUserData
+
+	response, err = json.Marshal(_response)
+	return
+}
+
+/*
+AuthorizePrivateChannel allows you to authorize a users subscription to a
+private channel. It returns an authorization signature to send back to the client
+and authorize them.
+
+For more information see our docs: http://pusher.com/docs/authorizing_users.
+
+This is an example of authorizing a private-channel, using the built-in
 Golang HTTP library to start a server.
 
 In order to authorize a client, one must read the response into type `[]byte`
@@ -478,7 +556,7 @@ to send back to the client.
 	func pusherAuth(res http.ResponseWriter, req *http.Request) {
 
 		params, _ := ioutil.ReadAll(req.Body)
-		response, err := client.AuthenticatePrivateChannel(params)
+		response, err := client.AuthorizePrivateChannel(params)
 		if err != nil {
 			panic(err)
 		}
@@ -491,13 +569,24 @@ to send back to the client.
 		http.ListenAndServe(":5000", nil)
 	}
 */
-func (c *Client) AuthenticatePrivateChannel(params []byte) (response []byte, err error) {
-	return c.authenticateChannel(params, nil)
+func (c *Client) AuthorizePrivateChannel(params []byte) (response []byte, err error) {
+	return c.authorizeChannel(params, nil)
 }
 
 /*
-AuthenticatePresenceChannel allows you to authenticate a users subscription to a
-presence channel. It returns authentication signature to send back to the client
+AuthenticatePrivateChannel allows you to authorize a users subscription to a
+private channel. It returns an authorization signature to send back to the client
+and authorize them.
+
+Deprecated: use AuthorizePrivateChannel instead.
+*/
+func (c *Client) AuthenticatePrivateChannel(params []byte) (response []byte, err error) {
+	return c.authorizeChannel(params, nil)
+}
+
+/*
+AuthorizePresenceChannel allows you to authorize a users subscription to a
+presence channel. It returns an authorization signature to send back to the client
 and authorize them. In order to identify a user, clients are sent a user_id and,
 optionally, custom data.
 
@@ -512,19 +601,31 @@ In this library, one does this by passing a `pusher.MemberData` instance.
 		},
 	}
 
-	response, err := client.AuthenticatePresenceChannel(params, presenceData)
+	response, err := client.AuthorizePresenceChannel(params, presenceData)
 	if err != nil {
 		panic(err)
 	}
 
 	fmt.Fprintf(res, response)
 */
-func (c *Client) AuthenticatePresenceChannel(params []byte, member MemberData) (response []byte, err error) {
-	return c.authenticateChannel(params, &member)
+func (c *Client) AuthorizePresenceChannel(params []byte, member MemberData) (response []byte, err error) {
+	return c.authorizeChannel(params, &member)
 }
 
-func (c *Client) authenticateChannel(params []byte, member *MemberData) (response []byte, err error) {
-	channelName, socketID, err := parseAuthRequestParams(params)
+/*
+AuthenticatePresenceChannel allows you to authorize a users subscription to a
+presence channel. It returns an authorization signature to send back to the client
+and authorize them. In order to identify a user, clients are sent a user_id and,
+optionally, custom data.
+
+Deprecated: use AuthorizePresenceChannel instead.
+*/
+func (c *Client) AuthenticatePresenceChannel(params []byte, member MemberData) (response []byte, err error) {
+	return c.authorizeChannel(params, &member)
+}
+
+func (c *Client) authorizeChannel(params []byte, member *MemberData) (response []byte, err error) {
+	channelName, socketID, err := parseChannelAuthorizationRequestParams(params)
 	if err != nil {
 		return
 	}